# swagger.json
{"openapi":"3.0.0","paths":{"/appointment/get/professional-appointments":{"get":{"operationId":"AppointmentController_getAppointment","summary":"Get the appointment data","parameters":[{"name":"patientId","required":false,"in":"query","schema":{"type":"string"}},{"name":"startDate","required":false,"in":"query","schema":{"type":"string"}},{"name":"endDate","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a JSON with the appointment data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAppointmentResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["appointment"],"security":[{"auth":[]}]}},"/appointment/get/my-appointments":{"get":{"operationId":"AppointmentController_getMyAppointments","summary":"Get all the client data appointments","parameters":[],"responses":{"200":{"description":"Returns a JSON with the data from all client data appointments","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAppointmentResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["appointment"],"security":[{"auth":[]}]}},"/appointment/post":{"post":{"operationId":"AppointmentController_postAppointment","summary":"Post the appointment data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostAppointmentReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the appointment data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAppointmentResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["appointment"],"security":[{"auth":[]}]}},"/appointment/put":{"put":{"operationId":"AppointmentController_putAppointment","summary":"Put the appointment data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutAppointmentReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the appointment data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAppointmentResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["appointment"],"security":[{"auth":[]}]}},"/appointment/patch/link-appointment/{uuid}":{"patch":{"operationId":"AppointmentController_patchLinkAppointment","summary":"Get all the client data appointments","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a JSON with the data from all client data appointments","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAppointmentResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["appointment"],"security":[{"auth":[]}]}},"/appointment/patch/cancel-appointment/{uuid}":{"patch":{"operationId":"AppointmentController_patchCancelAppointment","summary":"Patch the appointment data to cancel a patientId","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a JSON with the appointment status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAppointmentResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["appointment"],"security":[{"auth":[]}]}},"/appointment/delete/{uuid}":{"delete":{"operationId":"AppointmentController_deleteAppointment","summary":"Delete the appointment data","parameters":[{"name":"uuid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns a JSON with the appointment status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteAppointmentResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["appointment"],"security":[{"auth":[]}]}}},"info":{"title":"ingt-ms-appointment","description":"Rational Emotive Behaviour Therapy - Dysfunctional Thought Record","version":"1.0.0","contact":{}},"tags":[],"servers":[{"url":"/ingt-ms-appointment"}],"components":{"securitySchemes":{"auth":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","in":"header"}},"schemas":{"GetAppointmentResDto":{"type":"object","properties":{"appointment":{"type":"array","items":{"type":"string"}}},"required":["appointment"]},"ErrorDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"PostAppointmentReqDto":{"type":"object","properties":{"patientId":{"type":"string"},"date":{"type":"string"},"startTime":{"type":"string"},"endTime":{"type":"string"},"type":{"type":"string"},"location":{"type":"string"}},"required":["patientId","date","startTime","endTime","type","location"]},"PutAppointmentReqDto":{"type":"object","properties":{"uuid":{"type":"string"},"date":{"type":"string"},"startTime":{"type":"string"},"endTime":{"type":"string"},"type":{"type":"string"},"location":{"type":"string"}},"required":["uuid","date","startTime","endTime","type","location"]},"DeleteAppointmentResDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]}}}}